Backend/requirements.txt
boto3==1.18.28
botocore==1.21.28
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
Pillow==8.3.1
psycopg2-binary==2.9.1
python-dateutil==2.8.2
python-dotenv==0.19.0
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
SQLAlchemy==1.4.23
urllib3==1.26.6
Werkzeug==2.0.1
